Minor fix: squash functions & init_migrate incident DB
